% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Add.Assay}
\alias{Add.Assay}
\title{Add a Palette.Assay to a Palette object}
\usage{
Add.Assay(assay, Palette.object, check = TRUE, class = "original", name = NULL)
}
\arguments{
\item{assay}{A Palette.Assay.}

\item{Palette.object}{A Palette object.}

\item{check}{Whether to check the feature or cell information between \code{assay} and \code{Palette.object}.}

\item{class}{The data category carried by the assay, where ‘original’ represents real data and ‘infer’ represents inferred data。}

\item{name}{The name of \code{assay}.}
}
\value{
\code{Palette} object.
}
\description{
Add a Palette.Assay to a Palette object
}
\keyword{internal}
