% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brightness.R
\name{brightness}
\alias{brightness}
\alias{darken}
\alias{lighten}
\alias{darken.default}
\alias{darken.palette}
\alias{lighten.default}
\alias{lighten.palette}
\title{Control palette brightness}
\usage{
darken(x, amount = 0.1, ...)

lighten(x, amount = 0.1, ...)

\method{darken}{default}(x, amount = 0.1, ...)

\method{darken}{palette}(x, amount = 0.1, ...)

\method{lighten}{default}(x, amount = 0.1, ...)

\method{lighten}{palette}(x, amount = 0.1, ...)
}
\arguments{
\item{x}{A vector of colors}

\item{amount}{A numeric value to control the brightness of the palette}

\item{...}{Additional arguments passed to other methods. Currently ignored.}
}
\value{
A palette vector with the brightness adjusted
}
\description{
Generic function to control the brightness of a palette.
}
\examples{
darken(palette(roygbiv))
lighten(palette(roygbiv))
}
