% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Build.Kernel.cluster}
\alias{Build.Kernel.cluster}
\title{This function constructs a global kernel (similarity) matrix across batches or modalities by integrating cluster structures, using dimensionality reduction, L2 normalization, and cosine similarity–based filtering to identify reliable inter-cluster links.}
\usage{
Build.Kernel.cluster(
  data.list,
  cluster,
  dim.reduce = TRUE,
  cos.dims = NULL,
  Angle.var = 10,
  max.Angle = 35,
  joint = TRUE,
  seed = 123L
)
}
\arguments{
\item{data.list}{A list of matrices containing input data from different batches/modalities.}

\item{cluster}{A list of clustering assignments corresponding to \code{data.list}.}

\item{dim.reduce}{Logical; whether to perform dimensionality reduction before kernel construction.}

\item{cos.dims}{Dimensions to use for cosine similarity calculation.}

\item{Angle.var}{Angular variance (in degrees) used as a threshold for filtering weak links.}

\item{max.Angle}{Maximum angular threshold (in degrees) for acceptable inter-cluster similarity.}

\item{joint}{Logical; whether to perform joint pairwise dimensionality reduction across datasets.}

\item{seed}{Random seed to ensure reproducibility.}
}
\description{
This function constructs a global kernel (similarity) matrix across batches or modalities by integrating cluster structures, using dimensionality reduction, L2 normalization, and cosine similarity–based filtering to identify reliable inter-cluster links.
}
\keyword{internal}
