% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Build.Kernel.cluster_S}
\alias{Build.Kernel.cluster_S}
\title{Build simplified kernel matrix across clusters}
\usage{
Build.Kernel.cluster_S(data.list, cluster, seed = 123L)
}
\arguments{
\item{data.list}{A list of matrices containing input data from different batches or modalities.}

\item{cluster}{A list of clustering assignments corresponding to each dataset in \code{data.list}.}

\item{seed}{An integer random seed to ensure reproducibility (default: 123L).}
}
\value{
A list containing two sparse matrices:
\item{K}{The constructed global kernel (similarity) matrix.}
\item{B}{The corresponding block (batch-specific) kernel matrix.}
}
\description{
\code{Build.Kernel.cluster_S} constructs a simplified global kernel matrix (\code{K}) and
corresponding block matrix (\code{B}) across clusters, based on input data matrices
and cluster assignments, using normalized kernel construction.
}
\keyword{internal}
