% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Create.Palette.Assay}
\alias{Create.Palette.Assay}
\title{Create a new Palette.Assay class}
\usage{
Create.Palette.Assay(
  count,
  sample,
  modal,
  filter = FALSE,
  min.cells = 0,
  min.features = 0,
  slot = "Raw",
  sparse = TRUE
)
}
\arguments{
\item{count}{A data matrix with feature by cell.}

\item{sample}{A character represents the sample/batch information for the data matrix.}

\item{modal}{A character represents the modality information for the data matrix.}

\item{filter}{Whether to filter the data.}

\item{min.cells}{Include features detected in at least this many cells. Use only if \code{filter = TRUE}.}

\item{min.features}{Include cells where at least this many features are detected. Use only if \code{filter = TRUE}.}

\item{slot}{Which slot ('Raw' or 'data') to put the \code{count}.}

\item{sparse}{Whether to use sparse format ('dgCMatirx' class). (default TRUE)}
}
\value{
\code{Palette.Assay} object.
}
\description{
Create a new Palette.Assay class
}
\keyword{internal}
