% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Create.Palette.Object}
\alias{Create.Palette.Object}
\title{Create a new Palette object}
\usage{
Create.Palette.Object(
  data.list,
  samples,
  modals,
  filter = FALSE,
  min.cells = 0,
  min.features = 0,
  modal.order,
  sparse = TRUE
)
}
\arguments{
\item{data.list}{List of multimodal data matrix. If the names of the list are not \code{NULL}, then the names information will be used to name each Palette.Assay. If the names of the list are \code{NULL}, then the combined sample and modal information will be used to name each Palette.Assay.}

\item{samples}{A character vector of up to the same length as data.list, with each element represent the sample/batch information for each data matrix.}

\item{modals}{A character vector of up to the same length as data.list, with each element represent the modality information for each data matrix.}

\item{filter}{Whether to filter the data.}

\item{min.cells}{Include features detected in at least this many cells. Use only if \code{filter = TRUE}.}

\item{min.features}{Include cells where at least this many features are detected. Use only if \code{filter = TRUE}.}

\item{modal.order}{Orders of modalities for filter condition}

\item{sparse}{Whether use sparse format ('dgCMatirx' class). (default TRUE)}
}
\value{
\code{Palette} object.
}
\description{
Create a new Palette object
}
