% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Find.Cluster}
\alias{Find.Cluster}
\title{This function dentifies cell clusters within a Palette object using dimension reduction (PCA, LSI or SVD) and neighbor graph–based clustering, and optionally selects representative cells from each cluster based on silhouette scores and quantile thresholds.}
\usage{
Find.Cluster(
  object,
  modal = NULL,
  sample = NULL,
  dim.reduce = TRUE,
  dims = list(1:50),
  method = c("PCA", "LSI", "SVD"),
  joint = FALSE,
  resolution = 1,
  nn.k = 20,
  nn.method = "annoy",
  annoy.metric = "euclidean",
  prune.SNN = 1/15,
  downsample = TRUE,
  quantile.cutoff = 0.75,
  min.cell = 20,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Palette object containing assay data.}

\item{modal}{Modalities to be clustered (e.g., RNA, ATAC, ADT).}

\item{sample}{Specific sample(s) to include in clustering.}

\item{dim.reduce}{Whether to perform clustering on reduced-dimensional data.}

\item{dims}{Dimensions to use for clustering.}

\item{method}{Dimensionality reduction method ('PCA', 'LSI' or 'SVD').}

\item{joint}{Whether to cluster in a shared low-dimensional space across modalities.}

\item{resolution}{Resolution parameter controlling clustering granularity.}

\item{nn.k}{Number of neighbors for kNN/SNN graph construction.}

\item{nn.method}{Neighbor search method (default "annoy").}

\item{annoy.metric}{Distance metric for annoy-based neighbor search.}

\item{prune.SNN}{Pruning parameter for shared nearest-neighbor graph.}

\item{downsample}{Whether to select representative cells within each cluster.}

\item{quantile.cutoff}{Quantile threshold for selecting representative cells.}

\item{min.cell}{Minimum number of cells to retain per cluster.}

\item{verbose}{Whether to print progress messages.}
}
\description{
This function dentifies cell clusters within a Palette object using dimension reduction (PCA, LSI or SVD) and neighbor graph–based clustering, and optionally selects representative cells from each cluster based on silhouette scores and quantile thresholds.
}
