% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Get.Features}
\alias{Get.Features}
\title{Get the feature name information of a modality}
\usage{
Get.Features(object = NULL, modal = NULL, sample = NULL, check = FALSE)
}
\arguments{
\item{object}{A \code{Palette} object.}

\item{modal}{A character of modality.}

\item{sample}{A character of sample/batch.}

\item{check}{Whether to traverse all batches contained in the modality and take the feature union.}
}
\value{
\code{Palette} object.
}
\description{
Get the feature name information of a modality
}
\keyword{internal}
