% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{IDFLog.Data}
\alias{IDFLog.Data}
\title{This function performs inverse document frequency (IDF)–based scaling followed by log normalization on the raw assay data of a specified modality (and sample) within a Palette object.}
\usage{
IDFLog.Data(
  object,
  modal = NULL,
  sample = NULL,
  scale.factor = 10000,
  idf = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{Palette} object.}

\item{modal}{A character of modality.}

\item{sample}{A character of sample/batch.}

\item{scale.factor}{The scale factor used for each cell. (default 10000)}

\item{idf}{A precomputed IDF vector to use. If NULL, compute based on the input data matrix.}

\item{verbose}{Whether to show progress bar for calculations}
}
\value{
\code{Palette} object.
}
\description{
This function performs inverse document frequency (IDF)–based scaling followed by log normalization on the raw assay data of a specified modality (and sample) within a Palette object.
}
