% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{list_trc_main_colours}
\alias{list_trc_main_colours}
\title{List Main TRC Colors with Names}
\usage{
list_trc_main_colours()
}
\value{
A data frame with two columns: \code{name} (color name) and \code{hex} (hex code).
}
\description{
This function returns a data frame that lists the names of the main TRC colors and their corresponding hex codes.
}
\examples{
list_trc_main_colours()
}
