% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Normalize.Data}
\alias{Normalize.Data}
\title{Normalize the count data present in a given assay.}
\usage{
Normalize.Data(
  object,
  modal = NULL,
  normal.method = "LogNormalize",
  scale.factor = 10000,
  margin = 1
)
}
\arguments{
\item{object}{A \code{Palette} object.}

\item{modal}{A character vector of modality names to be normalized.}

\item{normal.method}{Method for normalization.('LogNormalize' for transcriptome, and 'CLR' for proteome)}

\item{scale.factor}{The scale factor used for each cell. (default 10000)}

\item{margin}{If performing CLR normalization, normalize across features (1) or cells (2)}
}
\value{
\code{Palette} object.
}
\description{
Normalize the count data present in a given assay.
}
