% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\docType{class}
\name{Palette-class}
\alias{Palette-class}
\alias{Palette}
\title{The Palette Object}
\description{
The Palette Object is a S4 classes for single-cell multimodal data
and associated information, such as dimensionality reduction embeddings,
mosaic integration results.
}
\section{Slots}{

\describe{
\item{\code{Assays}}{List of Palette.Assay.}

\item{\code{Meta}}{A data.frame containing meta informatoin for each Palette.Assay.}

\item{\code{Data.index}}{List of information for each Palette.Assay, including modality and sample/batch information.}

\item{\code{Co.Features}}{Identified high variable features shared across batches.}

\item{\code{pre.clusters}}{List of clustering information for each data matrix.}

\item{\code{Int.result}}{List of integrated low-dimensional embedding.}
}}

