% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-package.R
\docType{package}
\name{palette-package}
\alias{palette}
\alias{palette-package}
\title{palette: Color Scheme Helpers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Hexadecimal codes are typically used to represent colors in R. Connecting these codes to their colors requires practice or memorization. 'palette' provides a 'vctrs' class for working with color palettes, including printing and plotting functions. The goal of the class is to place visual representations of color palettes directly on or, at least, next to their corresponding character representations. Palette extensions also are provided for data frames using 'pillar'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/palette}
  \item Report bugs at \url{https://github.com/christopherkenny/palette/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
