% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\docType{class}
\name{Palette.Assay-class}
\alias{Palette.Assay-class}
\alias{Palette.Assay}
\title{The Palette.Assay Class}
\description{
The Palette.Assay class is an intermediate data storage class that
stores a data matrix and other related information needed for
performing downstream analyses.
}
\section{Slots}{

\describe{
\item{\code{Modal}}{Data modality information, such as RNA, ATAC.}

\item{\code{Sample}}{Data sample/batch information.}

\item{\code{Raw}}{Raw data matrix (features by cells).}

\item{\code{data}}{Normalized data matrix (features by cells).}

\item{\code{RepresentativeCell}}{List of information for representative cells, including cell names and cluster infromation.}

\item{\code{embedding}}{Intra-modal joint dimensionality reduction results.}
}}

