% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette}
\alias{palette}
\alias{validate_palette}
\alias{is_palette}
\alias{as_palette}
\title{Create a palette}
\usage{
palette(x = character())

validate_palette(x)

is_palette(x)

as_palette(x)
}
\arguments{
\item{x}{A character vector of colors}
}
\value{
a palette vector
}
\description{
Create a palette
}
\examples{
palette(roygbiv)
}
