% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{palette_browse}
\alias{palette_browse}
\alias{palette_browse_url}
\title{Browse a palette on coolors.co}
\usage{
palette_browse(x)

palette_browse_url(x)
}
\arguments{
\item{x}{A palette}
}
\value{
A URL to the palette
}
\description{
Browse a palette on coolors.co
}
\examples{
palette_browse(roygbiv)
}
