% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{palette_decode_url}
\alias{palette_decode_url}
\title{Convert a URL to a palette}
\usage{
palette_decode_url(x)
}
\arguments{
\item{x}{A URL from either 'coolors.co' or 'colorhunt.co'}
}
\value{
a vector of type \code{palette}
}
\description{
This function takes a URL from either <coolors.co> or <colorhunt.co> and
returns the palette.
}
\examples{
palette_decode_url('https://coolors.co/ff4444-ffae4d-ffff60-50ff50-3939fa')
palette_decode_url('https://www.colorhunt.co/palette/b5c0d0ccd3caf5e8ddeed3d9')
}
