% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{RQ.Rep}
\alias{RQ.Rep}
\title{Get the inferred embedding of the query data (Reference-based integration).}
\usage{
RQ.Rep(
  reference = NULL,
  query = NULL,
  modal_ref = NULL,
  modal_query = NULL,
  batch_ref = NULL,
  batch_query = NULL,
  modal_order,
  method = "fastMNN",
  is.normalize = TRUE,
  normalize_method = c("LogNormalize", "CLR", "TF-IDF"),
  method_out.npcs = 40,
  latent,
  Bi_sPCA = TRUE,
  dims = 30,
  lambda = 0.5,
  do.scale = TRUE,
  do.center = TRUE,
  nn.k = 20,
  nn.method = "annoy",
  annoy.metric = "euclidean",
  prune.SNN = 1/15,
  verbose = TRUE
)
}
\arguments{
\item{reference}{A list of expression matrices or data frames for each reference sample.}

\item{query}{A list of expression matrices or data frames for each query sample.}

\item{modal_ref}{A character vector indicating the modality of each reference sample.}

\item{modal_query}{A character vector indicating the modality of each query sample.}

\item{batch_ref}{A character or factor vector indicating the batch of each reference sample.}

\item{batch_query}{A character or factor vector indicating the batch of each query sample.}

\item{modal_order}{Optional vector specifying the desired order of modalities for integration.}

\item{method}{Integration method for each modality. Options include 'fastMNN', 'Seurat', 'Harmony'.}

\item{is.normalize}{Logical or vector specifying whether to normalize each modality.}

\item{normalize_method}{Normalization method(s) for each modality. Options: 'LogNormalize','CLR','TF-IDF'.}

\item{method_out.npcs}{Number of principal components to output for each modality's integration.}

\item{latent}{A matrix representing latent features to guide the integration.}

\item{Bi_sPCA}{Logical indicating whether to use Bi-supervised Principal Component Analysis (Bi-sPCA) for subspace alignment.}

\item{dims}{Number of dimensions for the Bi-sPCA or subspace projection.}

\item{lambda}{Weighting parameter for Bi-sPCA subspace combination, between 0 and 1.}

\item{do.scale}{Logical indicating whether to scale data before subspace calculation.}

\item{do.center}{Logical indicating whether to center data before subspace calculation.}

\item{nn.k}{Number of nearest neighbors for constructing graphs in Bi-sPCA.}

\item{nn.method}{Method for nearest neighbor search. Default is 'annoy'.}

\item{annoy.metric}{Distance metric for Annoy. Default is 'euclidean'.}

\item{prune.SNN}{Threshold to prune shared nearest neighbor graph edges.}

\item{verbose}{Logical indicating whether to print progress messages.}
}
\value{
A list of representations for each modality composition block.
}
\description{
This function integrates and aligns datasets from different modalities and batches,
using specified integration methods and returns the representations for alignment.
}
