% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Run_Multi_SVD}
\alias{Run_Multi_SVD}
\title{This function performs multi-dataset singular value decomposition (SVD) to obtain a shared low-dimensional representation, using either covariance aggregation or concatenation strategies depending on data size.}
\usage{
Run_Multi_SVD(x, n.dims = 50, thresh = 10000L)
}
\arguments{
\item{x}{A list of matrices to be jointly decomposed.}

\item{n.dims}{Number of dimensions (singular vectors) to retain in the reduced representation.}

\item{thresh}{Row-size threshold determining whether to use standard covariance-based SVD or large-scale matrix methods.}
}
\description{
This function performs multi-dataset singular value decomposition (SVD) to obtain a shared low-dimensional representation, using either covariance aggregation or concatenation strategies depending on data size.
}
\keyword{internal}
