% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation.R
\name{saturation}
\alias{saturation}
\alias{saturate}
\alias{desaturate}
\alias{saturate.default}
\alias{saturate.palette}
\alias{desaturate.default}
\alias{desaturate.palette}
\title{Control palette saturation}
\usage{
saturate(x, amount = 0.1, ...)

desaturate(x, amount = 0.1, ...)

\method{saturate}{default}(x, amount = 0.1, ...)

\method{saturate}{palette}(x, amount = 0.1, ...)

\method{desaturate}{default}(x, amount = 0.1, ...)

\method{desaturate}{palette}(x, amount = 0.1, ...)
}
\arguments{
\item{x}{A vector of colors}

\item{amount}{A numeric value to control the saturation of the palette}

\item{...}{Additional arguments passed to other methods. Currently ignored.}
}
\value{
A palette vector with the saturation adjusted
}
\description{
Generic function to control the saturation of a palette.
}
\examples{
saturate(palette(roygbiv))
desaturate(palette(roygbiv))
}
