% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_fill_trc}
\alias{scale_fill_trc}
\title{Custom ggplot2 Fill Scale Using TRC Colors}
\usage{
scale_fill_trc(palette = trc_main_colours, reverse = FALSE, ...)
}
\arguments{
\item{palette}{The palette function to use (e.g., trc_main_colours)}

\item{reverse}{Should the palette be reversed? Default is FALSE.}

\item{...}{Additional arguments passed to ggplot2's scale_fill_manual}
}
\value{
A ggplot2 scale object for fills
}
\description{
This function applies the TRC main color palette to ggplot2 fills.
}
