% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{split_block}
\alias{split_block}
\title{Split modalities into blocks by batch}
\usage{
split_block(modal, batch)
}
\arguments{
\item{modal}{A character vector of modalities corresponding to samples.}

\item{batch}{A character or factor vector indicating the batch of each sample.}
}
\value{
A list where each element contains the indices of batches sharing the same set of modalities.
}
\description{
This function groups modalities by batch and identifies unique modality combinations across batches.
}
\keyword{internal}
