% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Trans.Knowledge}
\alias{Trans.Knowledge}
\title{Transfer labels from reference to query using k-NN}
\usage{
Trans.Knowledge(ref, query, ref.label = NULL, k = 5)
}
\arguments{
\item{ref}{A matrix or data frame of reference features (rows = samples, columns = features).}

\item{query}{A matrix or data frame of query features to be labeled.}

\item{ref.label}{A data frame or vector of labels corresponding to the reference samples.}

\item{k}{Integer. Number of nearest neighbors to use in the k-NN classifier.}
}
\value{
A data frame with predicted labels for the query samples.
}
\description{
This function predicts labels for a query dataset based on a reference dataset
using k-nearest neighbors classification.
}
