% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{trc_blues}
\alias{trc_blues}
\title{Blue Shades}
\usage{
trc_blues(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing blue shades.
}
\description{
This function provides a range of blue shades used in TRC visualizations.
}
\examples{
trc_blues()      # Default order
trc_blues(TRUE)  # Reversed order
}
