% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{trc_extended}
\alias{trc_extended}
\title{TRC Extended Main Colors}
\usage{
trc_extended(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing the extended main colors.
}
\description{
This function provides an extended color palette based on the main TRC colors.
}
\examples{
trc_extended()      # Default order
trc_extended(TRUE)  # Reversed order
}
