% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{trc_reds}
\alias{trc_reds}
\title{Red Shades}
\usage{
trc_reds(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing red shades.
}
\description{
This function provides a range of red shades used in TRC visualizations.
}
\examples{
trc_reds()      # Default order
trc_reds(TRUE)  # Reversed order
}
