% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Add.HVFs}
\alias{Add.HVFs}
\title{Specify a set of features as HVFs for a modality in Palette object.}
\usage{
Add.HVFs(object, modal = NULL, features = NULL)
}
\arguments{
\item{object}{A \code{Palette} object.}

\item{modal}{A character of modality.}

\item{features}{A character vector specifying the feature names to use as the HVF.If \code{features = NULL}, all shared features of the modality are used as HVFs.}
}
\value{
A character vector of feature names.
}
\description{
Specify a set of features as HVFs for a modality in Palette object.
}
