% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteSC.R
\name{Find.HVFs}
\alias{Find.HVFs}
\title{Identifies highly variable features (HVFs).}
\usage{
Find.HVFs(object, modal, nfeatures = 2000, verbose = FALSE)
}
\arguments{
\item{object}{A \code{Palette} object.}

\item{modal}{A character of modality.}

\item{nfeatures}{The number of selected features in each sample/batch.}

\item{verbose}{Whether to show progress bar for calculations}
}
\value{
\code{Palette} object.
}
\description{
Identifies highly variable features (HVFs).
}
