% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette}
\alias{palette}
\alias{validate_palette}
\alias{is_palette}
\alias{as_palette}
\title{Create a palette}
\usage{
palette(x = character())

validate_palette(x)

is_palette(x)

as_palette(x)
}
\arguments{
\item{x}{A character vector of colors}
}
\value{
a palette vector
}
\description{
A \code{palette} is a class for color palettes built on \code{vctrs}. It is represented
as a character vector of hexadecimal color codes. Named colors (like \code{'red'})
are translated to hexadecimal color codes.
}
\examples{
palette(roygbiv)
}
