% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_function.R
\name{palette_function}
\alias{palette_function}
\title{Build a palette function}
\usage{
palette_function(x)
}
\arguments{
\item{x}{a palette}
}
\value{
a function which takes an integer \code{n} and returns a vector of \code{n} colors
}
\description{
Creates a palette function for use within \code{ggplot2} as an argument to \code{discrete_scale}.
}
\examples{
palette_function(palette(roygbiv))(10)
}
