% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot Palette Colors}
\usage{
plot_palette(x)
}
\arguments{
\item{x}{A palette}
}
\value{
A \code{ggplot} (if ggplot2 is available) or base plot of the palette
}
\description{
Plot Palette Colors
}
\examples{
plot_palette(crayons$standard16)
}
