% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot Palette Colors}
\usage{
plot_palette(x, use_names = TRUE)
}
\arguments{
\item{x}{a palette}

\item{use_names}{Should the names of the palette be used as labels? Default: \code{TRUE}.}
}
\value{
A \code{ggplot} (if ggplot2 is available) or base plot of the palette
}
\description{
Plot Palette Colors
}
\examples{
plot_palette(c('#ED0A3F', '#0066FF', '#FBE870', '#01A638', '#FF681F'))
}
