% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{vec_ptype2.palette}
\alias{vec_ptype2.palette}
\title{Palette Coercion}
\usage{
\method{vec_ptype2}{palette}(x, y, ...)
}
\arguments{
\item{x, y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a vector of the same length, as class \code{palette} if convertible, otherwise \code{character}
}
\description{
Coercion methods for \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}
}
\keyword{internal}
